pyproject.toml
[project]
name = "sanchos-irish-tacos"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jupyter>=1.1.1",
    "logfire[fastapi]>=3.12.0",
    "pydantic-ai[examples]>=0.0.48",
    "python-dotenv>=1.1.0",
]
